requirements.txt
# Django and DRF
Django>=4.2,<5.1
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.0
psycopg>=3.1.0

# Environment variables
python-decouple>=3.8
dj-database-url>=2.1.0

# OpenAI API (also works with Groq)
openai>=1.0.0

# CORS
django-cors-headers>=4.3.0

# Image processing
Pillow>=10.0.0

# Rate limiting
django-ratelimit>=4.1.0

# Production server
gunicorn>=21.0.0

# Static files (WhiteNoise for serving static files)
whitenoise>=6.6.0
